refactor oderatin
